Problems
Note:
the json file I made: rpmoelle_cc1.json
the dat file I made: sDat.dat
the file to run to turn json to dat: cc_json_utils.py
Note: you have to enter the json and  dat filenames in the console

Current problems:
1)When I print out my .dat files, information it spits out is wrong: The
lower field does not print out at all, though a large blank is placed
where it should be. The hint, monster coordinates, and trap coordinates
(as well as the cloning machine coordinates) are wrong too.

2) When I load my dat file into tile world, all the objects appear as
expected, but when I enter a space next to a key, monster, hint, or
door, the character cannot move again. The grunt sound plays and the
sprite changes direction, but he doesn't actually move. The monsters
don't move either. Is there supposed to be something on the lower layer
that corresponds to this? I cannot find an example in the sample level
pack. I also can't move on to the next level. I can win level 1, but
then the game hangs. After I win level 1 and quit, I now always start
the game on level 2.